Completed VastHash

+++ src/nosmap.c
@@ -1,9 +1,7 @@
-#include "vasthash.h"
-#include "nosvec.h"
 #include "nosmap.h"
 
 
-NOSMap nosmap_new(uint64_t initial_capacity, uint32_t data_size, uint32_t data_type) {
+NOSMap nosmap_new(size_t initial_capacity, size_t data_size, uint32_t data_type) {
 	NOSMap map = {0, NULL};
 	map.buckets = nosvec_new(initial_capacity, data_size, data_type);
 	if (map.buckets.v == NULL) {
@@ -15,7 +13,7 @@ NOSMap nosmap_new(uint64_t initial_capacity, uint32_t data_size, uint32_t data_t
 }
 
 // NOSMapSearchResult nosmap__find_bucket(NOSMap *map, uint8_t *key) {
-// 	uint64_t index = vast_hash(key) % map->size;
+// 	size_t index = vast_hash(key) % map->size;
 // }
 
 #ifdef TESTS
