+++ src/nosmap.h
@@ -1,17 +1,20 @@
 #pragma once
+#include "vasthash.h"
+#include "nosvec.h"
 
 
 typedef struct {
-	uint64_t index;
+	size_t index;
 	bool found;
 } NOSMapSearchResult;
 
 typedef struct {
-	uint64_t size;
+	size_t size;
 	NOSVec buckets;
 } NOSMap;
 
-NOSMap nosmap_new(uint64_t initial_capacity, uint32_t data_size, uint32_t data_type);
+NOSMap nosmap_new(size_t initial_capacity, size_t data_size, uint32_t data_type);
+
 #ifdef TESTS
 void nosmap_test();
 #endif
